Re-organize I/O traits
Use naming conventions closer to tokio (into_split),
remove unnecessary code, move impls into their own module
