+++ crates/hring-bingo-ktls/src/main.rs
@@ -10,7 +10,7 @@ use std::{
 
 use color_eyre::eyre;
 use hring::{
-    buffet::{RollMut, SplitOwned},
+    buffet::{IntoSplit, RollMut},
     h1, h2,
     tokio_uring::net::TcpStream,
     Body, Encoder, ExpectResponseHeaders, Method, Request, Responder, ResponseDone, ServerDriver,
@@ -154,11 +154,11 @@ async fn handle_plaintext_conn(
     match proto {
         Proto::H1(h1_conf) => {
             info!("Using HTTP/1.1");
-            hring::h1::serve(stream.split_owned(), h1_conf, buf, driver).await?;
+            hring::h1::serve(stream.into_split(), h1_conf, buf, driver).await?;
         }
         Proto::H2(h2_conf) => {
             info!("Using HTTP/2");
-            hring::h2::serve(stream.split_owned(), h2_conf, buf, Rc::new(driver)).await?;
+            hring::h2::serve(stream.into_split(), h2_conf, buf, Rc::new(driver)).await?;
         }
     }
 
@@ -200,11 +200,11 @@ async fn handle_tls_conn(
     match alpn_proto.as_deref() {
         Some("h2") => {
             info!("Using HTTP/2");
-            hring::h2::serve(stream.split_owned(), h2_conf, buf, Rc::new(driver)).await?;
+            hring::h2::serve(stream.into_split(), h2_conf, buf, Rc::new(driver)).await?;
         }
         Some("http/1.1") | None => {
             info!("Using HTTP/1.1");
-            hring::h1::serve(stream.split_owned(), h1_conf, buf, driver).await?;
+            hring::h1::serve(stream.into_split(), h1_conf, buf, driver).await?;
         }
         Some(other) => return Err(eyre::eyre!("Unsupported ALPN protocol: {}", other)),
     }
@@ -234,7 +234,7 @@ impl ServerDriver for SDriver {
 
         req.version = Version::HTTP_11;
         let (transport, respond) =
-            h1::request(transport.split_owned(), req, req_body, driver).await?;
+            h1::request(transport.into_split(), req, req_body, driver).await?;
 
         // don't re-use transport for now
         drop(transport);
@@ -342,7 +342,7 @@ async fn sample_http_request() -> color_eyre::Result<()> {
         headers: Default::default(),
     };
 
-    let (transport, _) = h1::request(transport.split_owned(), req, &mut (), driver).await?;
+    let (transport, _) = h1::request(transport.into_split(), req, &mut (), driver).await?;
     // don't re-use transport for now
     drop(transport);
